refactor: Moved recipe input row to separate file
